requirements.txt
yfinance>=0.2.40
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
feedparser>=6.0.0
requests>=2.31.0
python-dateutil>=2.8.0
